hehe? + crash skytils command + ma own security manager + attonantion settings + crums test + blur gui is now a mixin = fix twitter white girls = macro improvments

+++ src/main/java/io/github/kingstefan26/stefans_util/core/setting/impl/SliderSetting.java
@@ -8,40 +8,42 @@ import io.github.kingstefan26.stefans_util.core.setting.general.SettingsCore;
 import java.util.function.Consumer;
 
 public class SliderSetting extends AbstractSetting {
-    float min;
+    double min;
 
-    float max;
+    double max;
 
-    public SliderSetting(String name, basicModule parentModule, Double deafultValue, float min, float max, Consumer<Object> callback) {
+    public SliderSetting(String name, basicModule parentModule, double deafultValue, double min, double max, Consumer<Object> callback) {
         super(name, parentModule, SettingsCore.type.slider, callback);
         this.min = min;
         this.max = max;
         this.ConfigObject = new configObject(name, parentModule.getName(), deafultValue);
         setValue(getValue());
     }
-    public SliderSetting(String name, basicModule parentModule, Double deafultValue, int min, int max, Consumer<Object> callback, String comment) {
+
+    public SliderSetting(String name, basicModule parentModule, double deafultValue, double min, double max, Consumer<Object> callback, String comment) {
         super(name, parentModule, SettingsCore.type.slider, callback);
         this.min = min;
         this.max = max;
         this.comment = comment;
         this.ConfigObject = new configObject(name, parentModule.getName(), deafultValue);
     }
-    public Double getValue(){
+
+    public double getValue() {
         return this.ConfigObject.getDoubleValue();
     }
 
-    public void setValue(Double value){
-        if(!(value >= max)){
+    public void setValue(double value) {
+        if (!(value >= max)) {
             this.ConfigObject.setDoubleValue(value);
             this.callback.accept(value);
         }
     }
 
-    public float getMin() {
+    public double getMin() {
         return min;
     }
 
-    public float getMax() {
+    public double getMax() {
         return max;
     }
 }
